check if file exists before sym link

+++ install.sh
@@ -3,7 +3,9 @@
 # TODO use generic install method
 sudo apt install imagemagick
 # Create sym link, so this is visible to startup
-sudo ln -s $(which convert) /usr/local/bin/convert
+if ! [ -f /usr/local/bin/convert ]; then
+    sudo ln -s $(which convert) /usr/local/bin/convert
+fi
 
 # Update permissions
 chmod 777 ./set_wallpaper.sh
